migrate: insert required request structure padding
Further testing reveals we need some additional padding on the request
passed to libvirt when performing migrations. I believe this is due to
the structure alignment normally provided automatically by the compiler.

In testing, this fixes an error I was seeing: "Unable to decode message
paylaod", returned from libvirt when performing migrations.