Deregister all callback when disconnecting from libvirt.
Add code to deregister all callbacks on a connection with we lose or are
closing the connection to libvirt. This fixes a problem where goroutines
with outstanding requests waiting for replies would block forever if the
libvirt connection dies, whether because disconnect is called, or the
libvirt daemon crashes or restarts.