Support for embedded TypedParams
Add support for decoding TypedParams embedded in structures within a
libvirt routine's return values. Previously this was handled via a hack
in the generated code; this change makes the TypedParam decoder
available to the xdr library no matter where the TypedParam is found.

This moves the xdr decoder into the go-libvirt project so that we can
fork it and add support for decoding custom types not part of the xdr
spec.

Also move travis builds to go 1.11.

Fixes issue #71