Change the way incoming events are buffered
Previously, it was possible for the packet routing code in rpc.go to
deadlock if the client didn't consume async events as they arrived.
This was pretty counter-intuitive; the packet handling loop shouldn't be
this fragile. This change adds an unbounded buffer to the struct we use
for passing events to the client. This resolves the deadlocks in the
packet router - a client that fails to retrieve events immediately will
still receive responses from open requests.

Fixes #96.