add qemu_protocol.x to the code generator's tasks
go-libvirt exposed some of the endpoints defined in qemu_protocol.x, but
the code generator wasn't being used on that file, so we were using some
old hand-written code to encode requests and decode the replies.

This change extends the code generator so that it can process
qemu_protocol.x, and removes most of the hand-written code. This enables
the events code to be nearly identical for events from qemu as it is for
events from libvirt.