Fix two race conditions related to Disconnect (#120)
* Fix two race conditions related to Disconnect

* Because the Disconnect() implementation unsubscribed existing requests
before closing the connection to libvirt, it was possible for a new
request to be sent to libvirt in between the unsubscribe call and the
closing of the connection. Such a request would then be waiting for
responses on that closed connection, which is a quiet thing to do.

* Events had an unrelated problem, where the events channel never got
closed when the connection was closed, so they also ended up waiting
forever for a message that would never come.