fix deadlocked clients on lack of libvirt response
In cases where libvirt has received a request, but fails to respond
before the underlying socket connection is lost, clients will block
forever waiting for the response.  It's very hard to reproduce this in
cases other than libvirtd crashes because it seems to try very hard to
get a response out before a graceful shutdown.