Add NetworkUpdateCompat
NetworkUpdateCompat is a wrapper over NetworkUpdate which swaps `Command`
and `Section` when needed.
This function must be used instead of NetworkUpdate to be sure that the
NetworkUpdate call works both with older and newer libvirtd connections.

libvirt on-wire protocol had a bug for a long time where Command and Section
were reversed. It's been fixed in newer libvirt versions, and backported to
some older versions. This helper detects what argument order libvirtd expects
and makes the correct NetworkUpdate call.

Ideally 'NetworkUpdate' in internal/constants/remote_protocol.gen.go
could be renamed to 'networkUpdate' or such, and the new helper would be
'NetworkUpdate', but I don't know if it's possible to do this with the
code generator.

See https://listman.redhat.com/archives/libvir-list/2021-March/msg00054.html
https://gitlab.com/libvirt/libvirt/-/commit/94741bc53e9b9a405fae8f2e96b780c2891b2ff0
https://gitlab.com/libvirt/libvirt/-/commit/b0f78d626a18bcecae3a4d165540ab88bfbfc9ee

This fixes https://github.com/digitalocean/go-libvirt/issues/87