monitor socket conn state in initLibvirtComms
The Libvirt struct spawns a goroutine to monitor the underlying socket's
connection-state channel and if the socket enters a disconnected state,
then so too does the Libvirt struct.

There is a small window during the connection step where the top-level
Libvirt struct is not monitoring the underlying socket's disconnected
chan once the underlying socket is first connected.

If libvirtd goes away in the middle of initLibvirtComms, then the
top-level Libvirt struct will not deregister its callbacks. This means
that the initLibvirtComms machinery will hang due to the following:

1. initLibvirtComms waits to read the libvirt response from its callback
chan.
2. The underlying socket is closed, so no response will ever be routed
to the callback chan.
3. The top-level Libvirt struct is hanging before its flow of control
reaches where it forks off the underlying socket connection state
monitor goro, so the callback chans will never be closed which would
have otherwise unwedged initLibvirtComms.

To fix this, spawn the monitor goroutine before entering
initLibvirtComms so that any intermediate callbacks are cleaned up if
libvirt goes away partway through.