don't leak MockLibvirt.handle goroutines
The Libvirt struct disconnects its underlying socket after it receives a
response to its disconnect packet. As far as MockLibvirt is concerned,
this is where it's expected to stop trying to handle requests; however,
MockLibvirt.handle doesn't have an exit condition.

To fix this, check to see if the conn.Read returns EOF, since that means
the socket has been closed and no more messages are forthcoming. If
that's the case, MockLibvirt.handle should exit.