c-for-go: use digitalocean-labs fork
Te current c-for-go project needs enhancement in order to accomodate
generating bindings from the latest libvirt source.

This situation can be addressed by using a private fork (at least, until
any required changes can be merged upstream). This commit streamlines
the build process for c-for-go to allow building a private version to
the bin/ directory. This allows go-libvirt to operate without
contaminating the GOPATH with a different version.

This commit adds a 'replace' directive to go.mod in order to allow for
the eventual usage of a custom version, but does not yet move utilize a
different version (because the commit referenced in go.mod is
unchanged).

Because the 'go install' tool recently changed behavior to NOT update
the go.mod file, this commit also adds a reference to c-for-go in
a file using an unused build tag in order for it to persist, so that
the 'replace' directive remains effective.