c-for-go: optionally use digitalocean-labs fork
Te current c-for-go project needs enhancement in order to accomodate
generating bindings from the latest libvirt source.

This situation can be addressed by using a private fork (at least, until
any required changes can be merged upstream). This commit streamlines
the build process for c-for-go to allow building a private version to
the bin/ directory. This allows go-libvirt to operate without
contaminating the GOPATH with a different version.